updated and rewritten code for fetching and displaying posts, with infinite scroll and sorting by desc/asc
