+++ feed/index.html
@@ -414,10 +414,12 @@
           <div id="sort-menu" class="hidden relative flex flex-col items-end transition-all">
             <div
               class="absolute bg-darkGray flex flex-col items-center gap-2 w-40 mt-3 rounded-lg py-2 px-2 text-white text-sm">
-              <!-- Sort by newest -->
+              <!-- Sort by descending -->
               <button
+                id="descending"
                 class="px-6 py-2 w-full text-center rounded-lg hover:bg-eden-400 hover:text-black active:ring-1 active:ring-eden-400 flex items-center justify-start gap-2">
                 <svg
+                  id="descending-icon"
                   xmlns="http://www.w3.org/2000/svg"
                   fill="none"
                   aria-hidden="true"
@@ -427,12 +429,14 @@
                   class="size-4">
                   <path stroke-linecap="round" stroke-linejoin="round" d="m4.5 12.75 6 6 9-13.5" />
                 </svg>
-                <span type="button">Newest</span>
+                <span type="button">Descending</span>
               </button>
-              <!-- Sort by latest -->
+              <!-- Sort by ascending -->
               <button
+                id="ascending"
                 class="px-6 py-2 w-full text-center rounded-lg hover:bg-eden-400 hover:text-black active:ring-1 active:ring-eden-400 flex items-center justify-start gap-2">
                 <svg
+                  id="ascending-icon"
                   xmlns="http://www.w3.org/2000/svg"
                   fill="none"
                   aria-hidden="true"
@@ -442,7 +446,7 @@
                   class="size-4 opacity-0">
                   <path stroke-linecap="round" stroke-linejoin="round" d="m4.5 12.75 6 6 9-13.5" />
                 </svg>
-                <span type="button">Latest</span>
+                <span type="button">Ascending</span>
               </button>
             </div>
           </div>
@@ -451,10 +455,8 @@
       <!-- Feed -->
       <section class="px-3 mt-8">
         <!-- post -->
-
         <div id="feed-posts" class="flex flex-col mt-4 max-w-xl mx-auto">
           <div id="feed-info-message"></div>
-
           <!-- <div class="flex flex-col gap-4">
             <a href="#" class="flex gap-2 items-center w-fit -mb-2">
               <img
@@ -518,72 +520,8 @@
             aria-hidden="true"
             class="w-full border-b border-black border-opacity-10 dark:border-white dark:border-opacity-10 pt-4"></div>-->
         </div>
-        <!-- post -->
-        <!-- <div id="feed-posts" class="flex flex-col mt-4 max-w-xl mx-auto"> -->
-        <!-- <div id="post-container" class="flex flex-col gap-4">
-            <a href="#" class="flex gap-2 items-center w-fit -mb-2">
-              <img
-                class="w-8 h-8 sm:w-10 sm:h-10 rounded-full object-cover"
-                src="../src/assets/images/profile/autumn-goodman-vTL_qy03D1I-unsplash.webp"
-                alt="Profile image" />
-              <div class="flex flex-col">
-                <p class="text-sm">Jane Smith</p>
-                <p class="text-xs font-light opacity-80" aria-label="When content was posted">1h ago</p>
-              </div>
-            </a>
-            <div class="flex max-h-48 xs:max-h-60 sm:max-h-72 md:max-h-80 lg:max-h-96">
-              <img
-                class="object-cover rounded-lg"
-                src="../src/assets/images/profile/christopher-campbell-rDEOVtE7vOs-unsplash.webp"
-                alt="lorem ipsum" />
-            </div>
-            <h2 id="post-title">Post title</h2>
-            <p id="post-body" class="text-sm max-w-prose">
-              Lorem ipsum dolor sit amet consectetur adipisicing elit. Quae culpa, saepe deserunt iure
-              similique quo, ipsum est voluptates amet, pariatur illum? Quas voluptates nihil molestias,
-              temporibus incidunt ullam quidem dolorum!
-            </p>
-            <div class="flex gap-4">
-              <div>
-                <input
-                  type="checkbox"
-                  id="like"
-                  class="peer absolute size-6 icon-states opacity-0 cursor-pointer" />
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke-width="1.5"
-                  stroke="currentColor"
-                  class="size-6 icon-states peer-checked:fill-rose-600 peer-checked:stroke-rose-600 peer-hover:stroke-2 peer-focus-visible:outline-none peer-focus-visible:ring-2 peer-focus-visible:ring-black peer-focus-visible:rounded-sm">
-                  <path
-                    stroke-linecap="round"
-                    stroke-linejoin="round"
-                    d="M21 8.25c0-2.485-2.099-4.5-4.688-4.5-1.935 0-3.597 1.126-4.312 2.733-.715-1.607-2.377-2.733-4.313-2.733C5.1 3.75 3 5.765 3 8.25c0 7.22 9 12 9 12s9-4.78 9-12Z" />
-                </svg>
-              </div>
-              <button type="button">
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke-width="1.5"
-                  stroke="currentColor"
-                  class="size-6 icon-states">
-                  <path
-                    stroke-linecap="round"
-                    stroke-linejoin="round"
-                    d="M7.5 8.25h9m-9 3H12m-9.75 1.51c0 1.6 1.123 2.994 2.707 3.227 1.129.166 2.27.293 3.423.379.35.026.67.21.865.501L12 21l2.755-4.133a1.14 1.14 0 0 1 .865-.501 48.172 48.172 0 0 0 3.423-.379c1.584-.233 2.707-1.626 2.707-3.228V6.741c0-1.602-1.123-2.995-2.707-3.228A48.394 48.394 0 0 0 12 3c-2.392 0-4.744.175-7.043.513C3.373 3.746 2.25 5.14 2.25 6.741v6.018Z" />
-                </svg>
-              </button>
-            </div>
-          </div>
-          <div
-            aria-hidden="true"
-            class="w-full border-b border-black border-opacity-10 dark:border-white dark:border-opacity-10 pt-4"></div> -->
-        <!-- </div> -->
         <!-- loader -->
-        <div class="text-center">
+        <div id="loader" class="text-center">
           <div role="status">
             <svg
               aria-hidden="true"
