+++ src/css/styles.css
@@ -1004,10 +1004,6 @@ video {
   height: 100vh;
 }
 
-.h-12 {
-  height: 3rem;
-}
-
 .max-h-48 {
   max-height: 12rem;
 }
@@ -1022,6 +1018,10 @@ video {
 
 .w-2 {
   width: 0.5rem;
+}
+
+.w-20 {
+  width: 5rem;
 }
 
 .w-28 {
@@ -1047,14 +1047,6 @@ video {
 
 .w-screen {
   width: 100vw;
-}
-
-.w-12 {
-  width: 3rem;
-}
-
-.w-20 {
-  width: 5rem;
 }
 
 .max-w-prose {
@@ -1246,16 +1238,12 @@ video {
   fill: currentColor;
 }
 
-.fill-white {
-  fill: #F9F9F9;
-}
-
-.fill-blue-600 {
-  fill: #2563eb;
-}
-
 .fill-eden-400 {
   fill: #2AD7B2;
+}
+
+.fill-white {
+  fill: #F9F9F9;
 }
 
 .stroke-current {
@@ -1338,16 +1326,8 @@ video {
   padding-top: 1rem;
 }
 
-.text-left {
-  text-align: left;
-}
-
 .text-center {
   text-align: center;
-}
-
-.text-right {
-  text-align: right;
 }
 
 .text-2xl {
@@ -1383,6 +1363,11 @@ video {
   color: rgb(24 26 28 / var(--tw-text-opacity));
 }
 
+.text-eden-100 {
+  --tw-text-opacity: 1;
+  color: rgb(203 252 238 / var(--tw-text-opacity));
+}
+
 .text-eden-400 {
   --tw-text-opacity: 1;
   color: rgb(42 215 178 / var(--tw-text-opacity));
@@ -1391,26 +1376,6 @@ video {
 .text-white {
   --tw-text-opacity: 1;
   color: rgb(249 249 249 / var(--tw-text-opacity));
-}
-
-.text-gray-200 {
-  --tw-text-opacity: 1;
-  color: rgb(229 231 235 / var(--tw-text-opacity));
-}
-
-.text-gray-500 {
-  --tw-text-opacity: 1;
-  color: rgb(107 114 128 / var(--tw-text-opacity));
-}
-
-.text-gray-300 {
-  --tw-text-opacity: 1;
-  color: rgb(209 213 219 / var(--tw-text-opacity));
-}
-
-.text-eden-100 {
-  --tw-text-opacity: 1;
-  color: rgb(203 252 238 / var(--tw-text-opacity));
 }
 
 .underline {
@@ -1818,14 +1783,6 @@ input:-webkit-autofill,
   }
 }
 
-.rtl\:text-left:where([dir="rtl"], [dir="rtl"] *) {
-  text-align: left;
-}
-
-.rtl\:text-right:where([dir="rtl"], [dir="rtl"] *) {
-  text-align: right;
-}
-
 .dark\:border-eden-500:where(.dark, .dark *) {
   --tw-border-opacity: 1;
   border-color: rgb(17 187 153 / var(--tw-border-opacity));
@@ -1854,6 +1811,11 @@ input:-webkit-autofill,
   stroke: #11BB99;
 }
 
+.dark\:text-eden-100:where(.dark, .dark *) {
+  --tw-text-opacity: 1;
+  color: rgb(203 252 238 / var(--tw-text-opacity));
+}
+
 .dark\:text-eden-300:where(.dark, .dark *) {
   --tw-text-opacity: 1;
   color: rgb(91 237 201 / var(--tw-text-opacity));
@@ -1867,26 +1829,6 @@ input:-webkit-autofill,
 .dark\:text-white:where(.dark, .dark *) {
   --tw-text-opacity: 1;
   color: rgb(249 249 249 / var(--tw-text-opacity));
-}
-
-.dark\:text-gray-600:where(.dark, .dark *) {
-  --tw-text-opacity: 1;
-  color: rgb(75 85 99 / var(--tw-text-opacity));
-}
-
-.dark\:text-red-300:where(.dark, .dark *) {
-  --tw-text-opacity: 1;
-  color: rgb(252 165 165 / var(--tw-text-opacity));
-}
-
-.dark\:text-eden-600:where(.dark, .dark *) {
-  --tw-text-opacity: 1;
-  color: rgb(10 151 125 / var(--tw-text-opacity));
-}
-
-.dark\:text-eden-100:where(.dark, .dark *) {
-  --tw-text-opacity: 1;
-  color: rgb(203 252 238 / var(--tw-text-opacity));
 }
 
 .dark\:placeholder-white:where(.dark, .dark *)::-moz-placeholder {
