+++ src/js/api/posts/fetchPosts.mjs
@@ -1,10 +1,9 @@
 import { API_BASE_URL, API_SOCIAL_ENDPOINT, API_POSTS_ENDPOINT } from "../../constants/api.mjs";
 import { displayMessage } from "../../ui/common/displayMessage.mjs";
-import { isFieldsetDisabled } from "../../ui/common/utils/isFieldsetDisabled.mjs";
 import { createAPIRequestHeader } from "../utils/createAPIRequestHeader.mjs";
 
-export const getPosts = async () => {
-  const createPostURL = `${API_BASE_URL}${API_SOCIAL_ENDPOINT}${API_POSTS_ENDPOINT}`;
+export const fetchPosts = async (param) => {
+  const createPostURL = `${API_BASE_URL}${API_SOCIAL_ENDPOINT}${API_POSTS_ENDPOINT}?${param}`;
   const form = document.querySelector("#create-post-form");
   const messageContainer = document.querySelector("#info-message");
