+++ src/js/constants/api.mjs
@@ -4,5 +4,4 @@ export const API_REGISTER_ENDPOINT = "/register";
 export const API_LOGIN_ENDPOINT = "/login";
 export const API_SOCIAL_ENDPOINT = "/social";
 export const API_POSTS_ENDPOINT = "/posts";
-
 export const API_KEY = "75b5c809-1e1b-4ea0-8942-da0c692f881f";
