+++ src/js/events/posts/getTimeAgo.mjs
@@ -1,0 +1,12 @@
+export const getTimeAgo = (created) => {
+  const currentDate = new Date();
+  const postCreated = new Date(created);
+  const diff = Math.floor((currentDate - postCreated) / 1000); // Difference in seconds
+
+  if (diff < 60) return `${diff} seconds ago`;
+  if (diff < 3600) return `${Math.floor(diff / 60)}min ago`;
+  if (diff < 86400) return `${Math.floor(diff / 3600)}h ago`;
+  if (diff < 2592000) return `${Math.floor(diff / 86400)}d ago`;
+  if (diff < 31536000) return `${Math.floor(diff / 2592000)}mo ago`;
+  return `${Math.floor(diff / 31536000)}y ago`;
+};
